Fixed cart bug, made cart ui better
Fixed a bug where the cart would not open after an item was installed. Furthermore, the cart UI was not clear enough to see items clearly. A lot of whitespaces. White spaces were removed, order price area was changed and pictures are to show from now onwards for every item which is added